movie_maniac/requirements.txt
Django==1.7
django-tastypie==0.12.1
lxml==3.4.4
MySQL-python==1.2.5
python-dateutil==2.4.2
python-mimeparse==0.1.4
six==1.9.0
gunicorn==19.1.1
wsgiref==0.1.2
dj-database-url==0.3.0
dj-static==0.0.6
psycopg2==2.6.1
